requirements.txt
openai>=1.6.1
pinecone>=3.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1
langgraph>=0.0.20
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3 